Move generic constants, dataclasses, and helpers into nemoguardrails/tracing.py
